+++ src/leiningen/new/re_frame/src/cljs/core.cljs
@@ -1,6 +1,7 @@
 (ns {{ns-name}}.core
   (:require
    [reagent.core :as reagent]
+   [reagent.dom :as rdom]
    [re-frame.core :as re-frame]{{#re-pressed?}}
    [re-pressed.core :as rp]{{/re-pressed?}}{{#breaking-point?}}
    [breaking-point.core :as bp]{{/breaking-point?}}
@@ -17,8 +18,9 @@
 
 (defn ^:dev/after-load mount-root []
   (re-frame/clear-subscription-cache!)
-  (reagent/render [views/main-panel]
-                  (.getElementById js/document "app")))
+  (let [root-el (.getElementById js/document "app")]
+    (rdom/unmount-component-at-node root-el)
+    (rdom/render [views/main-panel] root-el)))
 
 (defn init []{{#routes?}}
   (routes/app-routes){{/routes?}}